requirements.txt
python-telegram-bot==21.3
google-generativeai==0.7.2
flask==3.0.3
gunicorn==22.0.0
aiohttp==3.9.5
Pillow==10.4.0
